python/NuixServerTest/pyproject.toml
[project]
name = "nuix-server-test"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "euxcet", email = "zcc.qwer@gmail.com"},
]
dependencies = [
    "requests>=2.31.0",
    "websockets>=12.0",
    "pytest>=8.2.0",
    "sseclient-py>=1.8.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
